% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetSim.R
\name{gen_graphs}
\alias{gen_graphs}
\title{This fucntion allowes you to generate three different types of networks with same number of vertices and edges}
\usage{
gen_graphs(ScaleFreePowerRange, SmallWorldProbability, VerticesVector,
  SampleSize, edgesEachScaleFree, savingDir, plotGraph = "y")
}
\arguments{
\item{ScaleFreePowerRange}{A vector containing powers for scale free networks, e.g. c(1.5,1.75)}

\item{SmallWorldProbability}{A vector containing probabilities for small world networks, e.g. c(0.5,0.6)}

\item{VerticesVector}{A vector containing number of vertices to be used to generate all three network types(please note that number of edges will be calculated to match number of edges and verices for all three network types, see documentation for further information), e.g. c(10,100)}

\item{SampleSize}{Integer number defining how many sample will be generated for each network types}

\item{savingDir}{A string with the directory where to save graph objects and plots, the directory folder need to be created beforehand}

\item{plotGraph}{if "y" then plots will also be created, only line plots to generate point plots use}

\item{edgesEachScaleFreeInteger}{number defining how many edges to be added in each step of scale free networks}
}
\description{
This fucntion allowes you to generate three different types of networks with same number of vertices and edges
}

