% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetSim.R
\name{gen_plots}
\alias{gen_plots}
\title{This function allowes you to generate three different types of networks with same number of vertices and edges}
\usage{
gen_plots(ScaleFreePowerRange, SmallWorldProbability, plot = C("point",
  "line"), savingDir)
}
\arguments{
\item{ScaleFreePowerRange}{A vector containing powers for scale free networks, e.g. c(1.5,1.75)}

\item{SmallWorldProbability}{A vector containing probabilities for small world networks, e.g. c(0.5,0.6)}

\item{plot}{type of plot(string input "point" or "line")}

\item{savingDir}{A string with the directory where to save graph plots. This function requires the network proporties to be generated and saved in that folder by using gen_graps function}
}
\description{
This function allowes you to generate three different types of networks with same number of vertices and edges
}

