\name{create_process_state}
\alias{add_network}
\alias{add_attribute_container}
\docType{package}
\title{
Manage the process state
}
\description{
A set of functions to manage the process state of the simulation Markov process.
The process state can be created and networks, attribute containers and global attributes be added.
The IDs of these objects can be requested and are necessary for some sub-sequent functions. 
}
\usage{
rcpp_hello_world()	
}
\examples{
\dontrun{
rcpp_hello_world()
}
}